seperate config case and pon tests
