+++ tests/commands/config/test_config_sample.py
@@ -92,7 +92,7 @@ def test_tumor_only_config(invoke_cli, sample_fastq, tmp_path,
 
 def test_dag_graph_success(tumor_normal_wgs_config, tumor_only_config,
                            tumor_normal_config, tumor_only_wgs_config,
-                           tumor_only_umi_config, pon_config):
+                           tumor_only_umi_config):
     # WHEN creating config using standard CLI input and setting Sentieon env vars
     # THEN DAG graph should be created successfully
     assert Path(json.load(
@@ -104,8 +104,6 @@ def test_dag_graph_success(tumor_normal_wgs_config, tumor_only_config,
         open(tumor_normal_wgs_config))["analysis"]["dag"]).exists()
     assert Path(json.load(
         open(tumor_only_umi_config))["analysis"]["dag"]).exists()
-    assert Path(json.load(
-        open(pon_config))["analysis"]["dag"]).exists()
 
 
 def test_config_bad_filename(
@@ -118,14 +116,10 @@ def test_config_bad_filename(
     # GIVEN existing fastq file with wrong naming convention
     faulty_fastq_dir = tmp_path_factory.mktemp("error_fastq")
     fastq_file_name_tumor = "tumor_error.fastq.gz"
-    fastq_file_name_normal = "normal_error.fastq.gz"
     Path(faulty_fastq_dir / fastq_file_name_tumor).touch()
-    Path(faulty_fastq_dir / fastq_file_name_normal).touch()
 
     case_id1 = "faulty_tumor"
-    case_id2 = "faulty_pon_normal"
     tumor = Path(faulty_fastq_dir / fastq_file_name_tumor).as_posix()
-    normal = Path(faulty_fastq_dir / fastq_file_name_normal).as_posix()
 
     # Invoke CLI command using file as argument
     case_result = invoke_cli(
@@ -145,26 +139,9 @@ def test_config_bad_filename(
         ],
     )
 
-    pon_result = invoke_cli(
-        [
-            "config",
-            "pon",
-            "-n",
-            normal,
-            "-p",
-            panel_bed_file,
-            "--case-id",
-            case_id2,
-            "--analysis-dir",
-            analysis_dir,
-            "--balsamic-cache",
-            balsamic_cache,
-        ],
-    )
 
     # THEN run should abort
     assert case_result.exit_code == 1
-    assert pon_result.exit_code == 1
 
 
 def test_run_without_permissions(
@@ -227,14 +204,10 @@ def test_config_graph_failed(invoke_cli, sample_fastq,
                                   panel_bed_file):
     # GIVEN an analysis config 
     case_id = "sample_tumor_only"
-    pon_case_id = "sample_pon"
     tumor = sample_fastq["tumor"]
-    normal = sample_fastq["normal"]
-    normal1 = sample_fastq["normal1"]
-    normal2 = sample_fastq["normal2"]
 
-    with mock.patch.object(graphviz, 'Source') as mocked1:
-        mocked1.return_value = None
+    with mock.patch.object(graphviz, 'Source') as mocked:
+        mocked.return_value = None
         case_result = invoke_cli(
             [
                 "config",
@@ -251,93 +224,6 @@ def test_config_graph_failed(invoke_cli, sample_fastq,
                 balsamic_cache
             ],
         )
-        with mock.patch.object(graphviz, 'Source') as mocked2:
-            mocked2.return_value = None
-            pon_result = invoke_cli(
-            [
-                "config",
-                "pon",
-                "-p",
-                panel_bed_file,
-                "-n",
-                normal,
-                "-n",
-                normal1,
-                "-n",
-                normal2,
-                "--case-id",
-                pon_case_id,
-                "--analysis-dir",
-                analysis_dir,
-                "--balsamic-cache",
-                balsamic_cache
-            ],
-        )
 
     assert case_result.exit_code == 1
-    assert pon_result.exit_code == 1
 
-
-def test_pon_config(invoke_cli, sample_fastq, tmp_path,
-                    balsamic_cache, panel_bed_file):
-    # GIVEN a case ID, fastq files, and an analysis dir
-    test_analysis_dir = tmp_path / "test_analysis_dir"
-    test_analysis_dir.mkdir()
-    case_id = "sample_pon"
-    normal = sample_fastq["normal"]
-    normal1 = sample_fastq["normal1"]
-    normal2 = sample_fastq["normal2"]
-
-    # WHEN creating a case config
-    result = invoke_cli(
-        [
-            "config",
-            "pon",
-            "--case-id",
-            case_id,
-            "-p",
-            panel_bed_file,
-            "-n",
-            normal,
-            "-n",
-            normal1,
-            "-n",
-            normal2,
-            "--analysis-dir",
-            test_analysis_dir,
-            "--balsamic-cache",
-            balsamic_cache
-            ]
-        )
-
-    # THEN a config should be created and exist
-    assert result.exit_code == 0
-    assert Path(test_analysis_dir, case_id, case_id + "_PON.json").exists()
-
-
-def test_pon_config_failed(invoke_cli, tmp_path,
-                    balsamic_cache, panel_bed_file):
-    # GIVEN a case ID, fastq files, and an analysis dir
-    test_analysis_dir = tmp_path / "test_analysis_dir"
-    test_analysis_dir.mkdir()
-    case_id = "sample_pon"
-
-    # WHEN creating a case analysis
-    result = invoke_cli(
-        [
-            "config",
-            "pon",
-            "--case-id",
-            case_id,
-            "-p",
-            panel_bed_file,
-            "--analysis-dir",
-            test_analysis_dir,
-            "--balsamic-cache",
-            balsamic_cache
-        ]
-    )
-
-    # THEN a config should be created and exist
-    assert 'Error: Missing option' in result.output
-    assert result.exit_code == 2
